[feat]: Record Pagination 구축 및 API 추가
